build: establish warning-free content baseline

+++ src/main/java/de/bixilon/minosoft/data/container/stack/properties/DurabilityProperty.kt
@@ -33,7 +33,7 @@ data class DurabilityProperty(
 
     override fun writeNbt(item: Item, version: Version, registries: Registries, nbt: MutableJsonObject) {
         val max = item.unsafeCast<DurableItem>().maxDurability
-        nbt[DAMAGE_TAG] = (max - durability).toInt()
+        nbt[DAMAGE_TAG] = max - durability
 
         if (unbreakable) {
             nbt[UNBREAKABLE_TAG] = 1.toByte()
